Deploy cluster agent for clusters with private-endpoint
In the case that the cluster agent could not be deployed by Rancher, a
bastion is used to deployed the cluster agent. An annotation is used to
detect when the bastion should be deployed and deleted. This annotation
is set and deleted by the Rancher implementation.

Since this implementation has the possibility of costing a user money
with AWS, a flag/option may be added so a user could explicity disabled
this feature and run the import command manually.